run forge fmt

+++ src/infinity/facets/Bridges/C3/RetireC3C3TFacet.sol
@@ -16,7 +16,7 @@ contract RetireC3C3TFacet is ReentrancyGuard {
         string retirementMessage,
         address indexed carbonPool,
         address carbonToken,
-        uint retiredAmount
+        uint256 retiredAmount
     );
 
     /**
@@ -36,12 +36,12 @@ contract RetireC3C3TFacet is ReentrancyGuard {
      */
     function c3RetireExactC3T(
         address carbonToken,
-        uint amount,
+        uint256 amount,
         address beneficiaryAddress,
         string memory beneficiaryString,
         string memory retirementMessage,
         LibTransfer.From fromMode
-    ) external nonReentrant returns (uint retirementIndex) {
+    ) external nonReentrant returns (uint256 retirementIndex) {
         // Currently this is a simple wrapper for direct calls on specific TCO2 tokens
         // No fee is charged
 
@@ -75,13 +75,13 @@ contract RetireC3C3TFacet is ReentrancyGuard {
      */
     function c3RetireExactC3TWithEntity(
         address carbonToken,
-        uint amount,
+        uint256 amount,
         string memory retiringEntityString,
         address beneficiaryAddress,
         string memory beneficiaryString,
         string memory retirementMessage,
         LibTransfer.From fromMode
-    ) external nonReentrant returns (uint retirementIndex) {
+    ) external nonReentrant returns (uint256 retirementIndex) {
         // Currently this is a simple wrapper for direct calls on specific TCO2 tokens
         // No fee is charged
