+++ src/infinity/libraries/LibMeta.sol
@@ -5,22 +5,19 @@ library LibMeta {
     bytes32 internal constant EIP712_DOMAIN_TYPEHASH =
         keccak256(bytes("EIP712Domain(string name,string version,uint256 salt,address verifyingContract)"));
 
-    function domainSeparator(
-        string memory name,
-        string memory version
-    ) internal view returns (bytes32 domainSeparator_) {
+    function domainSeparator(string memory name, string memory version)
+        internal
+        view
+        returns (bytes32 domainSeparator_)
+    {
         domainSeparator_ = keccak256(
             abi.encode(
-                EIP712_DOMAIN_TYPEHASH,
-                keccak256(bytes(name)),
-                keccak256(bytes(version)),
-                getChainID(),
-                address(this)
+                EIP712_DOMAIN_TYPEHASH, keccak256(bytes(name)), keccak256(bytes(version)), getChainID(), address(this)
             )
         );
     }
 
-    function getChainID() internal view returns (uint id) {
+    function getChainID() internal view returns (uint256 id) {
         assembly {
             id := chainid()
         }
@@ -29,7 +26,7 @@ library LibMeta {
     function msgSender() internal view returns (address sender_) {
         if (msg.sender == address(this)) {
             bytes memory array = msg.data;
-            uint index = msg.data.length;
+            uint256 index = msg.data.length;
             assembly {
                 // Load the 32 bytes word from memory with the address on the lower 20 bytes, and mask those.
                 sender_ := and(mload(add(array, index)), 0xffffffffffffffffffffffffffffffffffffffff)
